Added more

+++ src/main/java/org/block/plugin/PluginContainer.java
@@ -13,12 +13,15 @@ import org.block.util.MapBuilder;
 import org.block.util.ReflectProcessor;
 
 import java.io.File;
-import java.lang.annotation.Annotation;
 import java.net.URISyntaxException;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
+/**
+ * This provides the standard data for every plugin launch. Every plugin will have one and should be
+ * used for if a plugin is required as a return or a parameter.
+ */
 public final class PluginContainer implements Comparable<PluginContainer>{
 
     private Object plugin;
@@ -41,10 +44,23 @@ public final class PluginContainer implements Comparable<PluginContainer>{
         OPEN_BLOCK_CONTAINER = container;
     }
 
+    /**
+     * Init
+     * @param plugin The main of the plugin
+     * @param filePath The file path to the plugin
+     * @throws IllegalArgumentException If the "plugin" does not contain @Plugin
+     */
     public PluginContainer(Object plugin, File filePath){
         this(plugin, filePath, false);
     }
 
+    /**
+     * Init
+     * Provided "isDisabled" for plugins that are specified within the plugins folder however are disabled
+     * @param plugin The main of the plugin
+     * @param filePath The file path to the public
+     * @param isDisabled If the plugin is disabled
+     */
     public PluginContainer(Object plugin, File filePath, boolean isDisabled){
         if(!plugin.getClass().isAnnotationPresent(Plugin.class)){
             throw new IllegalArgumentException("The provided plugin does not contain @Plugin");
@@ -54,6 +70,10 @@ public final class PluginContainer implements Comparable<PluginContainer>{
         this.filePath = filePath;
     }
 
+    /**
+     * Gets the modules from the plugin
+     * @return The module from the plugins
+     */
     public Set<Module> getModules(){
         if(this.modules.isEmpty()){
             ModuleRegisterEvent event = new ModuleRegisterEvent();
@@ -63,31 +83,61 @@ public final class PluginContainer implements Comparable<PluginContainer>{
         return this.modules;
     }
 
+    /**
+     * Checks if the plugin is disabled
+     * @return If the plugin is disabled
+     */
     public boolean isDisabled(){
         return this.isDisabled;
     }
 
+    /**
+     * Gets the main of the plugin
+     * @return The mains instance
+     */
     public Object getPlugin(){
         return this.plugin;
     }
 
+    /**
+     * Gets the @Plugin from the mains instance
+     * @return The mains instance
+     */
     public Plugin getPluginMeta(){
         return this.plugin.getClass().getAnnotation(Plugin.class);
     }
 
+    /**
+     * Gets the location of the plugin
+     * @return The location of the plugin
+     */
     public File getFilePath(){
         return this.filePath;
     }
 
+    /**
+     * Gets the plugins directory for plugin specific general files such as config files.
+     * These will be removed if the user uninstalls the plugin though the GUI with the option to remove config enabled
+     * @return The folder location of the plugins directory for general files
+     */
     public File getConfigDirectory(){
         return new File("config/" + this.getPluginMeta().id() + "/");
     }
 
-    public void delete(){
-        GeneralUntil.getFiles(this.getConfigDirectory()).parallelStream().forEach(f -> f.deleteOnExit());
+    /**
+     * Deletes the plugin
+     * @param removeConfigFiles If true, then all general files for the plugin will be removed
+     */
+    public void delete(boolean removeConfigFiles){
+        if(removeConfigFiles) {
+            GeneralUntil.getFiles(this.getConfigDirectory()).parallelStream().forEach(f -> f.deleteOnExit());
+        }
         this.getFilePath().deleteOnExit();
     }
 
+    /**
+     * This is used for the plugin loader, ignore
+     */
     public void initInjects(){
         initInjects(new MapBuilder.MapArrayBuilder<InjectData, Object>()
                 .put(InjectData.PLUGIN_CONTAINER, this)
@@ -96,6 +146,10 @@ public final class PluginContainer implements Comparable<PluginContainer>{
                 .build());
     }
 
+    /**
+     * This is used for the plugin loader, ignore
+     * @param data
+     */
     public void initInjects(Map<InjectData, Object[]> data){
         ReflectProcessor.getDirect(c -> c.getFields(), this.plugin).stream().filter(f -> f.isAnnotationPresent(Inject.class)).forEach(f -> {
             data.entrySet().parallelStream().forEach(entry -> {
@@ -111,9 +165,14 @@ public final class PluginContainer implements Comparable<PluginContainer>{
                 }
             });
         });
-
     }
 
+    /**
+     * Gives a order to all PluginContainers
+     * Most plugins will be of equal value, however those that depend on other plugins will be at higher value
+     * @param o The pluginContainer to compare
+     * @return The compared value
+     */
     @Override
     public int compareTo(PluginContainer o) {
         for(Dependent dependent : this.getPluginMeta().dependsOn()){
